:recycle: Refactor: domain 패키지 밑으로 리팩토링
